requirements.txt
fastapi
uvicorn
watchdog
torch
torchvision
transformers
pillow
python-multipart
langchain-huggingface
langchain-community
llama-cpp-python